requirements.txt
fastmcp>=0.1.0
pymongo>=4.6.2
python-dotenv>=1.0.0
requests>=2.31.0
langchain-text-splitters>=0.0.1
